Factor out project loading from MSBuildWorkspace
