+++ src/Workspaces/Core/Desktop/PublicAPI.txt
@@ -7,6 +7,18 @@ Microsoft.CodeAnalysis.Host.Mef.DesktopMefHostServices
 Microsoft.CodeAnalysis.Host.Mef.MefV1HostServices
 Microsoft.CodeAnalysis.Host.Mef.MefV1HostServices.GetExports<TExtension, TMetadata>() -> System.Collections.Generic.IEnumerable<System.Lazy<TExtension, TMetadata>>
 Microsoft.CodeAnalysis.Host.Mef.MefV1HostServices.GetExports<TExtension>() -> System.Collections.Generic.IEnumerable<System.Lazy<TExtension>>
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.AssociateFileExtensionWithLanguage(string projectFileExtension, string language) -> void
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.Clear() -> void
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.LoadMetadataForReferencedProjects.get -> bool
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.LoadMetadataForReferencedProjects.set -> void
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.LoadProjectAsync(string projectFilePath, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<Microsoft.CodeAnalysis.ProjectInfo>
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.LoadProjectsAsync(string projectFilePath, System.Collections.Generic.Dictionary<Microsoft.CodeAnalysis.ProjectId, Microsoft.CodeAnalysis.ProjectInfo> loadedProjects, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<Microsoft.CodeAnalysis.ProjectId>
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.LoadSolutionAsync(string solutionFilePath, System.Threading.CancellationToken cancellationToken = default(System.Threading.CancellationToken)) -> System.Threading.Tasks.Task<Microsoft.CodeAnalysis.SolutionInfo>
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.MSBuildProjectLoader(Microsoft.CodeAnalysis.Workspace workspace, System.Collections.Immutable.ImmutableDictionary<string, string> properties = null) -> void
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.Properties.get -> System.Collections.Immutable.ImmutableDictionary<string, string>
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.SkipUnrecognizedProjects.get -> bool
+Microsoft.CodeAnalysis.MSBuild.MSBuildProjectLoader.SkipUnrecognizedProjects.set -> void
 Microsoft.CodeAnalysis.MSBuild.MSBuildWorkspace
 Microsoft.CodeAnalysis.MSBuild.MSBuildWorkspace.AssociateFileExtensionWithLanguage(string projectFileExtension, string language) -> void
 Microsoft.CodeAnalysis.MSBuild.MSBuildWorkspace.CloseSolution() -> void
