+++ src/Workspaces/Core/Desktop/Workspace/MSBuild/MSBuildWorkspace.cs
@@ -32,26 +32,14 @@ namespace Microsoft.CodeAnalysis.MSBuild
         // used to serialize access to public methods
         private readonly NonReentrantLock _serializationLock = new NonReentrantLock();
 
-        // used to protect access to mutable state
-        private readonly NonReentrantLock _dataGuard = new NonReentrantLock();
-
-        private readonly Dictionary<string, string> _extensionToLanguageMap = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-        private readonly Dictionary<string, ProjectId> _projectPathToProjectIdMap = new Dictionary<string, ProjectId>(StringComparer.OrdinalIgnoreCase);
-        private readonly Dictionary<string, IProjectFileLoader> _projectPathToLoaderMap = new Dictionary<string, IProjectFileLoader>(StringComparer.OrdinalIgnoreCase);
-
-        private string _solutionFilePath;
-        private ImmutableDictionary<string, string> _properties;
+        private MSBuildProjectLoader _loader;
 
         private MSBuildWorkspace(
             HostServices hostServices,
             ImmutableDictionary<string, string> properties)
             : base(hostServices, "MSBuildWorkspace")
         {
-            // always make a copy of these build properties (no mutation please!)
-            _properties = properties ?? ImmutableDictionary<string, string>.Empty;
-            this.SetSolutionProperties(solutionFilePath: null);
-            this.LoadMetadataForReferencedProjects = false;
-            this.SkipUnrecognizedProjects = true;
+            _loader = new MSBuildProjectLoader(this, properties);
         }
 
         /// <summary>
@@ -99,7 +87,7 @@ namespace Microsoft.CodeAnalysis.MSBuild
         /// </summary>
         public ImmutableDictionary<string, string> Properties
         {
-            get { return _properties; }
+            get { return _loader.Properties; }
         }
 
         /// <summary>
@@ -107,7 +95,11 @@ namespace Microsoft.CodeAnalysis.MSBuild
         /// If the referenced project is already opened, the metadata will not be loaded.
         /// If the metadata assembly cannot be found the referenced project will be opened instead.
         /// </summary>
-        public bool LoadMetadataForReferencedProjects { get; set; }
+        public bool LoadMetadataForReferencedProjects
+        {
+            get { return _loader.LoadMetadataForReferencedProjects; }
+            set { _loader.LoadMetadataForReferencedProjects = value; }
+        }
 
         /// <summary>
         /// Determines if unrecognized projects are skipped when solutions or projects are opened.
@@ -120,27 +112,18 @@ namespace Microsoft.CodeAnalysis.MSBuild
         /// 
         /// If unrecognized projects cannot be skipped a corresponding exception is thrown.
         /// </summary>
-        public bool SkipUnrecognizedProjects { get; set; }
+        public bool SkipUnrecognizedProjects
+        {
+            get { return _loader.SkipUnrecognizedProjects; }
+            set { _loader.SkipUnrecognizedProjects = value; }
+        }
 
         /// <summary>
         /// Associates a project file extension with a language name.
         /// </summary>
         public void AssociateFileExtensionWithLanguage(string projectFileExtension, string language)
         {
-            if (language == null)
-            {
-                throw new ArgumentNullException(nameof(language));
-            }
-
-            if (projectFileExtension == null)
-            {
-                throw new ArgumentNullException(nameof(projectFileExtension));
-            }
-
-            using (_dataGuard.DisposableWait())
-            {
-                _extensionToLanguageMap[projectFileExtension] = language;
-            }
+            _loader.AssociateFileExtensionWithLanguage(projectFileExtension, language);
         }
 
         /// <summary>
@@ -158,193 +141,8 @@ namespace Microsoft.CodeAnalysis.MSBuild
         {
             base.ClearSolutionData();
 
-            using (_dataGuard.DisposableWait())
-            {
-                this.SetSolutionProperties(solutionFilePath: null);
-
-                // clear project related data
-                _projectPathToProjectIdMap.Clear();
-                _projectPathToLoaderMap.Clear();
-            }
-        }
-
-        private const string SolutionDirProperty = "SolutionDir";
-
-        private void SetSolutionProperties(string solutionFilePath)
-        {
-            _solutionFilePath = solutionFilePath;
-
-            if (!string.IsNullOrEmpty(solutionFilePath))
-            {
-                // When MSBuild is building an individual project, it doesn't define $(SolutionDir).
-                // However when building an .sln file, or when working inside Visual Studio,
-                // $(SolutionDir) is defined to be the directory where the .sln file is located.
-                // Some projects out there rely on $(SolutionDir) being set (although the best practice is to
-                // use MSBuildProjectDirectory which is always defined).
-                if (!string.IsNullOrEmpty(solutionFilePath))
-                {
-                    string solutionDirectory = Path.GetDirectoryName(solutionFilePath);
-                    if (!solutionDirectory.EndsWith(@"\", StringComparison.Ordinal))
-                    {
-                        solutionDirectory += @"\";
-                    }
-
-                    if (Directory.Exists(solutionDirectory))
-                    {
-                        _properties = _properties.SetItem(SolutionDirProperty, solutionDirectory);
-                    }
-                }
-            }
-        }
-
-        private ProjectId GetProjectId(string fullProjectPath)
-        {
-            using (_dataGuard.DisposableWait())
-            {
-                ProjectId id;
-                _projectPathToProjectIdMap.TryGetValue(fullProjectPath, out id);
-                return id;
-            }
-        }
-
-        private ProjectId GetOrCreateProjectId(string fullProjectPath)
-        {
-            using (_dataGuard.DisposableWait())
-            {
-                ProjectId id;
-                if (!_projectPathToProjectIdMap.TryGetValue(fullProjectPath, out id))
-                {
-                    id = ProjectId.CreateNewId(debugName: fullProjectPath);
-                    _projectPathToProjectIdMap.Add(fullProjectPath, id);
-                }
-
-                return id;
-            }
-        }
-
-        private bool TryGetLoaderFromProjectPath(string projectFilePath, ReportMode mode, out IProjectFileLoader loader)
-        {
-            using (_dataGuard.DisposableWait())
-            {
-                // check to see if we already know the loader
-                if (!_projectPathToLoaderMap.TryGetValue(projectFilePath, out loader))
-                {
-                    // otherwise try to figure it out from extension
-                    var extension = Path.GetExtension(projectFilePath);
-                    if (extension.Length > 0 && extension[0] == '.')
-                    {
-                        extension = extension.Substring(1);
-                    }
-
-                    string language;
-                    if (_extensionToLanguageMap.TryGetValue(extension, out language))
-                    {
-                        if (this.Services.SupportedLanguages.Contains(language))
-                        {
-                            loader = this.Services.GetLanguageServices(language).GetService<IProjectFileLoader>();
-                        }
-                        else
-                        {
-                            this.ReportFailure(mode, string.Format(WorkspacesResources.CannotOpenProjectUnsupportedLanguage, projectFilePath, language));
-                            return false;
-                        }
-                    }
-                    else
-                    {
-                        loader = ProjectFileLoader.GetLoaderForProjectFileExtension(this, extension);
-
-                        if (loader == null)
-                        {
-                            this.ReportFailure(mode, string.Format(WorkspacesResources.CannotOpenProjectUnrecognizedFileExtension, projectFilePath, Path.GetExtension(projectFilePath)));
-                            return false;
-                        }
-                    }
-
-                    if (loader != null)
-                    {
-                        _projectPathToLoaderMap[projectFilePath] = loader;
-                    }
-                }
-
-                return loader != null;
-            }
-        }
-
-        private bool TryGetAbsoluteProjectPath(string path, string baseDirectory, ReportMode mode, out string absolutePath)
-        {
-            try
-            {
-                absolutePath = this.GetAbsolutePath(path, baseDirectory);
-            }
-            catch (Exception)
-            {
-                ReportFailure(mode, string.Format(WorkspacesResources.InvalidProjectFilePath, path));
-                absolutePath = null;
-                return false;
-            }
-
-            if (!File.Exists(absolutePath))
-            {
-                ReportFailure(
-                    mode,
-                    string.Format(WorkspacesResources.ProjectFileNotFound, absolutePath),
-                    msg => new FileNotFoundException(msg));
-                return false;
-            }
-
-            return true;
-        }
-
-        private string GetAbsoluteSolutionPath(string path, string baseDirectory)
-        {
-            string absolutePath;
-
-            try
-            {
-                absolutePath = GetAbsolutePath(path, baseDirectory);
-            }
-            catch (Exception)
-            {
-                throw new InvalidOperationException(string.Format(WorkspacesResources.InvalidSolutionFilePath, path));
-            }
-
-            if (!File.Exists(absolutePath))
-            {
-                throw new FileNotFoundException(string.Format(WorkspacesResources.SolutionFileNotFound, absolutePath));
-            }
-
-            return absolutePath;
-        }
-
-        private enum ReportMode
-        {
-            Throw,
-            Log,
-            Ignore
-        }
-
-        private void ReportFailure(ReportMode mode, string message, Func<string, Exception> createException = null)
-        {
-            switch (mode)
-            {
-                case ReportMode.Throw:
-                    if (createException != null)
-                    {
-                        throw createException(message);
-                    }
-                    else
-                    {
-                        throw new InvalidOperationException(message);
-                    }
-
-                case ReportMode.Log:
-                    this.OnWorkspaceFailed(new WorkspaceDiagnostic(WorkspaceDiagnosticKind.Failure, message));
-                    break;
-
-                case ReportMode.Ignore:
-                default:
-                    break;
-            }
+            // clear project related data
+            _loader.Clear();
         }
 
         private string GetAbsolutePath(string path, string baseDirectoryPath)
@@ -352,7 +150,7 @@ namespace Microsoft.CodeAnalysis.MSBuild
             return Path.GetFullPath(FileUtilities.ResolveRelativePath(path, baseDirectoryPath) ?? path);
         }
 
-        #region Open Solution & Project
+#region Open Solution & Project
         /// <summary>
         /// Open a solution file and all referenced projects.
         /// </summary>
@@ -365,130 +163,10 @@ namespace Microsoft.CodeAnalysis.MSBuild
 
             this.ClearSolution();
 
-            var absoluteSolutionPath = this.GetAbsoluteSolutionPath(solutionFilePath, Directory.GetCurrentDirectory());
-
-            using (_dataGuard.DisposableWait(cancellationToken))
-            {
-                this.SetSolutionProperties(absoluteSolutionPath);
-            }
-
-            VersionStamp version = default(VersionStamp);
-
-#if !MSBUILD12
-            Microsoft.Build.Construction.SolutionFile solutionFile = Microsoft.Build.Construction.SolutionFile.Parse(absoluteSolutionPath);
-            var reportMode = this.SkipUnrecognizedProjects ? ReportMode.Log : ReportMode.Throw;
-            var invalidProjects = new List<ProjectInSolution>();
-
-            // seed loaders from known project types
-            using (_dataGuard.DisposableWait(cancellationToken))
-            {
-                foreach (var project in solutionFile.ProjectsInOrder)
-                {
-                    if (project.ProjectType == SolutionProjectType.SolutionFolder)
-                    {
-                        continue;
-                    }
-
-                    var projectAbsolutePath = TryGetAbsolutePath(project.AbsolutePath, reportMode);
-                    if (projectAbsolutePath != null)
-                    {
-                        var extension = Path.GetExtension(projectAbsolutePath);
-                        if (extension.Length > 0 && extension[0] == '.')
-                        {
-                            extension = extension.Substring(1);
-                        }
-
-                        var loader = ProjectFileLoader.GetLoaderForProjectFileExtension(this, extension);
-                        if (loader != null)
-                        {
-                            _projectPathToLoaderMap[projectAbsolutePath] = loader;
-                        }
-                    }
-                    else
-                    {
-                        invalidProjects.Add(project);
-                    }
-                }
-            }
-
-            // a list to accumulate all the loaded projects
-            var loadedProjects = new Dictionary<ProjectId, ProjectInfo>();
-
-            // load all the projects
-            foreach (var project in solutionFile.ProjectsInOrder)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-
-                if (project.ProjectType != SolutionProjectType.SolutionFolder && !invalidProjects.Contains(project))
-                {
-                    var projectAbsolutePath = TryGetAbsolutePath(project.AbsolutePath, reportMode);
-                    if (projectAbsolutePath != null)
-                    {
-                        IProjectFileLoader loader;
-                        if (TryGetLoaderFromProjectPath(projectAbsolutePath, reportMode, out loader))
-                        {
-                            // projects get added to 'loadedProjects' as side-effect
-                            // never prefer metadata when loading solution, all projects get loaded if they can.
-                            var tmp = await GetOrLoadProjectAsync(projectAbsolutePath, loader, preferMetadata: false, loadedProjects: loadedProjects, cancellationToken: cancellationToken).ConfigureAwait(false);
-                        }
-                    }
-                }
-            }
-#else
-            SolutionFile solutionFile = null;
-
-            using (var reader = new StreamReader(absoluteSolutionPath))
-            {
-                version = VersionStamp.Create(File.GetLastWriteTimeUtc(absoluteSolutionPath));
-                var text = await reader.ReadToEndAsync().ConfigureAwait(false);
-                solutionFile = SolutionFile.Parse(new StringReader(text));
-            }
-
-            var solutionFolder = Path.GetDirectoryName(absoluteSolutionPath);
-
-            // seed loaders from known project types
-            using (_dataGuard.DisposableWait())
-            {
-                foreach (var projectBlock in solutionFile.ProjectBlocks)
-                {
-                    string absoluteProjectPath;
-                    if (TryGetAbsoluteProjectPath(projectBlock.ProjectPath, solutionFolder, ReportMode.Ignore, out absoluteProjectPath))
-                    {
-                        var loader = ProjectFileLoader.GetLoaderForProjectTypeGuid(this, projectBlock.ProjectTypeGuid);
-                        if (loader != null)
-                        {
-                            _projectPathToLoaderMap[absoluteProjectPath] = loader;
-                        }
-                    }
-                }
-            }
-
-            // a list to accumulate all the loaded projects
-            var loadedProjects = new Dictionary<ProjectId, ProjectInfo>();
-
-            var reportMode = this.SkipUnrecognizedProjects ? ReportMode.Log : ReportMode.Throw;
-
-            // load all the projects
-            foreach (var projectBlock in solutionFile.ProjectBlocks)
-            {
-                cancellationToken.ThrowIfCancellationRequested();
-
-                string absoluteProjectPath;
-                if (TryGetAbsoluteProjectPath(projectBlock.ProjectPath, solutionFolder, reportMode, out absoluteProjectPath))
-                {
-                    IProjectFileLoader loader;
-                    if (TryGetLoaderFromProjectPath(absoluteProjectPath, reportMode, out loader))
-                    { 
-                        // projects get added to 'loadedProjects' as side-effect
-                        // never prefer metadata when loading solution, all projects get loaded if they can.
-                        var tmp = await GetOrLoadProjectAsync(absoluteProjectPath, loader, preferMetadata: false, loadedProjects: loadedProjects, cancellationToken: cancellationToken).ConfigureAwait(false);
-                    }
-                }
-            }
-#endif
+            var solutionInfo = await _loader.LoadSolutionAsync(solutionFilePath, cancellationToken).ConfigureAwait(false);
 
             // construct workspace from loaded project infos
-            this.OnSolutionAdded(SolutionInfo.Create(SolutionId.CreateNewId(debugName: absoluteSolutionPath), version, absoluteSolutionPath, loadedProjects.Values));
+            this.OnSolutionAdded(solutionInfo);
 
             this.UpdateReferencesAfterAdd();
 
@@ -505,458 +183,20 @@ namespace Microsoft.CodeAnalysis.MSBuild
                 throw new ArgumentNullException(nameof(projectFilePath));
             }
 
-            string fullPath;
-            if (this.TryGetAbsoluteProjectPath(projectFilePath, Directory.GetCurrentDirectory(), ReportMode.Throw, out fullPath))
+            var loadedProjects = new Dictionary<ProjectId, ProjectInfo>();
+            var projectId = await _loader.LoadProjectsAsync(projectFilePath, loadedProjects, cancellationToken).ConfigureAwait(false);
+
+            // add projects to solution
+            foreach (var project in loadedProjects.Values)
             {
-                IProjectFileLoader loader;
-                if (this.TryGetLoaderFromProjectPath(projectFilePath, ReportMode.Throw, out loader))
-                {
-                    var loadedProjects = new Dictionary<ProjectId, ProjectInfo>();
-                    var projectId = await GetOrLoadProjectAsync(fullPath, loader, this.LoadMetadataForReferencedProjects, loadedProjects, cancellationToken).ConfigureAwait(false);
-
-                    // add projects to solution
-                    foreach (var project in loadedProjects.Values)
-                    {
-                        this.OnProjectAdded(project);
-                    }
-
-                    this.UpdateReferencesAfterAdd();
-
-                    return this.CurrentSolution.GetProject(projectId);
-                }
+                this.OnProjectAdded(project);
             }
 
-            // unreachable
-            return null;
+            this.UpdateReferencesAfterAdd();
+
+            return this.CurrentSolution.GetProject(projectId);
         }
 
-        private string TryGetAbsolutePath(string path, ReportMode mode)
-        {
-            try
-            {
-                path = Path.GetFullPath(path);
-            }
-            catch (Exception)
-            {
-                ReportFailure(mode, string.Format(WorkspacesResources.InvalidProjectFilePath, path));
-                return null;
-            }
-
-            if (!File.Exists(path))
-            {
-                ReportFailure(
-                    mode,
-                    string.Format(WorkspacesResources.ProjectFileNotFound, path),
-                    msg => new FileNotFoundException(msg));
-                return null;
-            }
-
-            return path;
-        }
-
-        private void UpdateReferencesAfterAdd()
-        {
-            using (_serializationLock.DisposableWait())
-            {
-                var oldSolution = this.CurrentSolution;
-                var newSolution = this.UpdateReferencesAfterAdd(oldSolution);
-
-                if (newSolution != oldSolution)
-                {
-                    newSolution = this.SetCurrentSolution(newSolution);
-                    var ignore = this.RaiseWorkspaceChangedEventAsync(WorkspaceChangeKind.SolutionChanged, oldSolution, newSolution);
-                }
-            }
-        }
-
-        // Updates all projects to properly reference other existing projects via project references instead of using references to built metadata.
-        private Solution UpdateReferencesAfterAdd(Solution solution)
-        {
-            // Build map from output assembly path to ProjectId
-            // Use explicit loop instead of ToDictionary so we don't throw if multiple projects have same output assembly path.
-            var outputAssemblyToProjectIdMap = new Dictionary<string, ProjectId>();
-            foreach (var p in solution.Projects)
-            {
-                if (!string.IsNullOrEmpty(p.OutputFilePath))
-                {
-                    outputAssemblyToProjectIdMap[p.OutputFilePath] = p.Id;
-                }
-            }
-
-            // now fix each project if necessary
-            foreach (var pid in solution.ProjectIds)
-            {
-                var project = solution.GetProject(pid);
-
-                // convert metadata references to project references if the metadata reference matches some project's output assembly.
-                foreach (var meta in project.MetadataReferences)
-                {
-                    var pemeta = meta as PortableExecutableReference;
-                    if (pemeta != null)
-                    {
-                        ProjectId matchingProjectId;
-
-                        // check both Display and FilePath. FilePath points to the actually bits, but Display should match output path if 
-                        // the metadata reference is shadow copied.
-                        if ((!string.IsNullOrEmpty(pemeta.Display) && outputAssemblyToProjectIdMap.TryGetValue(pemeta.Display, out matchingProjectId)) ||
-                            (!string.IsNullOrEmpty(pemeta.FilePath) && outputAssemblyToProjectIdMap.TryGetValue(pemeta.FilePath, out matchingProjectId)))
-                        {
-                            var newProjRef = new ProjectReference(matchingProjectId, pemeta.Properties.Aliases, pemeta.Properties.EmbedInteropTypes);
-
-                            if (!project.ProjectReferences.Contains(newProjRef))
-                            {
-                                project = project.WithProjectReferences(project.ProjectReferences.Concat(newProjRef));
-                            }
-
-                            project = project.WithMetadataReferences(project.MetadataReferences.Where(mr => mr != meta));
-                        }
-                    }
-                }
-
-                solution = project.Solution;
-            }
-
-            return solution;
-        }
-
-        private async Task<ProjectId> GetOrLoadProjectAsync(string projectFilePath, IProjectFileLoader loader, bool preferMetadata, Dictionary<ProjectId, ProjectInfo> loadedProjects, CancellationToken cancellationToken)
-        {
-            var projectId = GetProjectId(projectFilePath);
-            if (projectId == null)
-            {
-                projectId = await this.LoadProjectAsync(projectFilePath, loader, preferMetadata, loadedProjects, cancellationToken).ConfigureAwait(false);
-            }
-
-            return projectId;
-        }
-
-        private async Task<ProjectId> LoadProjectAsync(string projectFilePath, IProjectFileLoader loader, bool preferMetadata, Dictionary<ProjectId, ProjectInfo> loadedProjects, CancellationToken cancellationToken)
-        {
-            Debug.Assert(projectFilePath != null);
-            Debug.Assert(loader != null);
-
-            var projectId = this.GetOrCreateProjectId(projectFilePath);
-
-            var projectName = Path.GetFileNameWithoutExtension(projectFilePath);
-
-            var projectFile = await loader.LoadProjectFileAsync(projectFilePath, _properties, cancellationToken).ConfigureAwait(false);
-            var projectFileInfo = await projectFile.GetProjectFileInfoAsync(cancellationToken).ConfigureAwait(false);
-
-            var projectDirectory = Path.GetDirectoryName(projectFilePath);
-            var outputFilePath = projectFileInfo.OutputFilePath;
-            var outputDirectory = Path.GetDirectoryName(outputFilePath);
-
-            VersionStamp version;
-            if (!string.IsNullOrEmpty(projectFilePath) && File.Exists(projectFilePath))
-            {
-                version = VersionStamp.Create(File.GetLastWriteTimeUtc(projectFilePath));
-            }
-            else
-            {
-                version = VersionStamp.Create();
-            }
-
-            // translate information from command line args
-            var commandLineParser = this.Services.GetLanguageServices(loader.Language).GetService<ICommandLineParserService>();
-            var metadataService = this.Services.GetService<IMetadataService>();
-            var analyzerService = this.Services.GetService<IAnalyzerService>();
-
-            var commandLineArgs = commandLineParser.Parse(
-                arguments: projectFileInfo.CommandLineArgs,
-                baseDirectory: projectDirectory, 
-                isInteractive: false, 
-                sdkDirectory: RuntimeEnvironment.GetRuntimeDirectory());
-
-            var resolver = new MetadataFileReferenceResolver(commandLineArgs.ReferencePaths, commandLineArgs.BaseDirectory);
-            var metadataReferences = commandLineArgs.ResolveMetadataReferences(new AssemblyReferenceResolver(resolver, metadataService.GetProvider()));
-
-            var analyzerLoader = analyzerService.GetLoader();
-            foreach (var path in commandLineArgs.AnalyzerReferences.Select(r => r.FilePath))
-            {
-                analyzerLoader.AddDependencyLocation(path);
-            }
-
-            var analyzerReferences = commandLineArgs.ResolveAnalyzerReferences(analyzerLoader);
-
-            var defaultEncoding = commandLineArgs.Encoding;
-
-            // docs & additional docs
-            var docFileInfos = projectFileInfo.Documents.ToImmutableArrayOrEmpty();
-            var additionalDocFileInfos = projectFileInfo.AdditionalDocuments.ToImmutableArrayOrEmpty();
-
-            // check for duplicate documents
-            var allDocFileInfos = docFileInfos.AddRange(additionalDocFileInfos);
-            CheckDocuments(allDocFileInfos, projectFilePath, projectId);
-
-            var docs = new List<DocumentInfo>();
-            foreach (var docFileInfo in docFileInfos)
-            {
-                string name;
-                ImmutableArray<string> folders;
-                GetDocumentNameAndFolders(docFileInfo.LogicalPath, out name, out folders);
-
-                docs.Add(DocumentInfo.Create(
-                    DocumentId.CreateNewId(projectId, debugName: docFileInfo.FilePath),
-                    name,
-                    folders,
-                    projectFile.GetSourceCodeKind(docFileInfo.FilePath),
-                    new FileTextLoader(docFileInfo.FilePath, defaultEncoding),
-                    docFileInfo.FilePath,
-                    docFileInfo.IsGenerated));
-            }
-
-            var additionalDocs = new List<DocumentInfo>();
-            foreach (var docFileInfo in additionalDocFileInfos)
-            {
-                string name;
-                ImmutableArray<string> folders;
-                GetDocumentNameAndFolders(docFileInfo.LogicalPath, out name, out folders);
-
-                additionalDocs.Add(DocumentInfo.Create(
-                    DocumentId.CreateNewId(projectId, debugName: docFileInfo.FilePath),
-                    name,
-                    folders,
-                    SourceCodeKind.Regular,
-                    new FileTextLoader(docFileInfo.FilePath, defaultEncoding),
-                    docFileInfo.FilePath,
-                    docFileInfo.IsGenerated));
-            }
-
-            // project references
-            var resolvedReferences = await this.ResolveProjectReferencesAsync(
-                projectId, projectFilePath, projectFileInfo.ProjectReferences, preferMetadata, loadedProjects, cancellationToken).ConfigureAwait(false);
-
-            // add metadata references for project refs converted to metadata refs
-            metadataReferences = metadataReferences.Concat(resolvedReferences.MetadataReferences);
-
-            // if the project file loader couldn't figure out an assembly name, make one using the project's file path.
-            var assemblyName = commandLineArgs.CompilationName;
-            if (string.IsNullOrWhiteSpace(assemblyName))
-            {
-                assemblyName = Path.GetFileNameWithoutExtension(projectFilePath);
-
-                // if this is still unreasonable, use a fixed name.
-                if (string.IsNullOrWhiteSpace(assemblyName))
-                {
-                    assemblyName = "assembly";
-                }
-            }
-
-            // make sure that doc-comments at least get parsed.
-            var parseOptions = commandLineArgs.ParseOptions;
-            if (parseOptions.DocumentationMode == DocumentationMode.None)
-            {
-                parseOptions = parseOptions.WithDocumentationMode(DocumentationMode.Parse);
-            }
-
-            // add all the extra options that are really behavior overrides
-            var compOptions = commandLineArgs.CompilationOptions
-                    .WithXmlReferenceResolver(new XmlFileResolver(projectDirectory))
-                    .WithSourceReferenceResolver(new SourceFileResolver(ImmutableArray<string>.Empty, projectDirectory))
-                    .WithMetadataReferenceResolver(
-                        new AssemblyReferenceResolver(
-                            new MetadataFileReferenceResolver(ImmutableArray<string>.Empty, projectDirectory),
-                            MetadataFileReferenceProvider.Default))
-                    .WithStrongNameProvider(new DesktopStrongNameProvider(ImmutableArray.Create(projectDirectory, outputFilePath)))
-                    .WithAssemblyIdentityComparer(DesktopAssemblyIdentityComparer.Default);
-
-            loadedProjects.Add(
-                projectId,
-                ProjectInfo.Create(
-                    projectId,
-                    version,
-                    projectName,
-                    assemblyName,
-                    loader.Language,
-                    projectFilePath,
-                    outputFilePath,
-                    compilationOptions: compOptions,
-                    parseOptions: parseOptions,
-                    documents: docs,
-                    projectReferences: resolvedReferences.ProjectReferences,
-                    metadataReferences: metadataReferences,
-                    analyzerReferences: analyzerReferences,
-                    additionalDocuments: additionalDocs,
-                    isSubmission: false,
-                    hostObjectType: null));
-
-            return projectId;
-        }
-
-        private static readonly char[] s_directorySplitChars = new char[] { Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar };
-
-        private static void GetDocumentNameAndFolders(string logicalPath, out string name, out ImmutableArray<string> folders)
-        {
-            var pathNames = logicalPath.Split(s_directorySplitChars, StringSplitOptions.RemoveEmptyEntries);
-            if (pathNames.Length > 0)
-            {
-                if (pathNames.Length > 1)
-                {
-                    folders = pathNames.Take(pathNames.Length - 1).ToImmutableArray();
-                }
-                else
-                {
-                    folders = ImmutableArray.Create<string>();
-                }
-
-                name = pathNames[pathNames.Length - 1];
-            }
-            else
-            {
-                name = logicalPath;
-                folders = ImmutableArray.Create<string>();
-            }
-        }
-
-        private void CheckDocuments(IEnumerable<DocumentFileInfo> docs, string projectFilePath, ProjectId projectId)
-        {
-            var paths = new HashSet<string>();
-            foreach (var doc in docs)
-            {
-                if (paths.Contains(doc.FilePath))
-                {
-                    this.OnWorkspaceFailed(new ProjectDiagnostic(WorkspaceDiagnosticKind.Warning, string.Format(WorkspacesResources.DuplicateSourceFileInProject, doc.FilePath, projectFilePath), projectId));
-                }
-
-                paths.Add(doc.FilePath);
-            }
-        }
-
-        private class ResolvedReferences
-        {
-            public readonly List<ProjectReference> ProjectReferences = new List<ProjectReference>();
-            public readonly List<MetadataReference> MetadataReferences = new List<MetadataReference>();
-        }
-
-        private async Task<ResolvedReferences> ResolveProjectReferencesAsync(
-            ProjectId thisProjectId,
-            string thisProjectPath,
-            IReadOnlyList<ProjectFileReference> projectFileReferences,
-            bool preferMetadata,
-            Dictionary<ProjectId, ProjectInfo> loadedProjects,
-            CancellationToken cancellationToken)
-        {
-            var resolvedReferences = new ResolvedReferences();
-            var reportMode = this.SkipUnrecognizedProjects ? ReportMode.Log : ReportMode.Throw;
-
-            foreach (var projectFileReference in projectFileReferences)
-            {
-                string fullPath;
-
-                if (TryGetAbsoluteProjectPath(projectFileReference.Path, Path.GetDirectoryName(thisProjectPath), reportMode, out fullPath))
-                {
-                    // if the project is already loaded, then just reference the one we have
-                    var existingProjectId = this.GetProjectId(fullPath);
-                    if (existingProjectId != null)
-                    {
-                        resolvedReferences.ProjectReferences.Add(new ProjectReference(existingProjectId, projectFileReference.Aliases));
-                        continue;
-                    }
-
-                    IProjectFileLoader loader;
-                    TryGetLoaderFromProjectPath(fullPath, ReportMode.Ignore, out loader);
-
-                    // get metadata if preferred or if loader is unknown
-                    if (preferMetadata || loader == null)
-                    {
-                        var projectMetadata = await this.GetProjectMetadata(fullPath, projectFileReference.Aliases, _properties, cancellationToken).ConfigureAwait(false);
-                        if (projectMetadata != null)
-                        {
-                            resolvedReferences.MetadataReferences.Add(projectMetadata);
-                            continue;
-                        }
-                    }
-
-                    // must load, so we really need loader
-                    if (TryGetLoaderFromProjectPath(fullPath, reportMode, out loader))
-                    {
-                        // load the project
-                        var projectId = await this.GetOrLoadProjectAsync(fullPath, loader, preferMetadata, loadedProjects, cancellationToken).ConfigureAwait(false);
-
-                        // If that other project already has a reference on us, this will cause a circularity.
-                        // This check doesn't need to be in the "already loaded" path above, since in any circularity this path
-                        // must be taken at least once.
-                        if (ProjectAlreadyReferencesProject(loadedProjects, projectId, targetProject: thisProjectId))
-                        {
-                            // We'll try to make this metadata if we can
-                            var projectMetadata = await this.GetProjectMetadata(fullPath, projectFileReference.Aliases, _properties, cancellationToken).ConfigureAwait(false);
-                            if (projectMetadata != null)
-                            {
-                                resolvedReferences.MetadataReferences.Add(projectMetadata);
-                            }
-                            continue;
-                        }
-                        else
-                        {
-                            resolvedReferences.ProjectReferences.Add(new ProjectReference(projectId, projectFileReference.Aliases));
-                            continue;
-                        }
-                    }
-                }
-                else
-                {
-                    fullPath = projectFileReference.Path;
-                }
-
-                // cannot find metadata and project cannot be loaded, so leave a project reference to a non-existent project.
-                var id = this.GetOrCreateProjectId(fullPath);
-                resolvedReferences.ProjectReferences.Add(new ProjectReference(id, projectFileReference.Aliases));
-            }
-
-            return resolvedReferences;
-        }
-
-        /// <summary>
-        /// Returns true if the project identified by <paramref name="fromProject"/> has a reference (even indirectly)
-        /// on the project identified by <paramref name="targetProject"/>.
-        /// </summary>
-        private bool ProjectAlreadyReferencesProject(Dictionary<ProjectId, ProjectInfo> loadedProjects, ProjectId fromProject, ProjectId targetProject)
-        {
-            ProjectInfo info;
-
-            return loadedProjects.TryGetValue(fromProject, out info) && info.ProjectReferences.Any(pr => pr.ProjectId == targetProject ||
-                ProjectAlreadyReferencesProject(loadedProjects, pr.ProjectId, targetProject));
-        }
-
-        /// <summary>
-        /// Gets a MetadataReference to a project's output assembly.
-        /// </summary>
-        private async Task<MetadataReference> GetProjectMetadata(string projectFilePath, ImmutableArray<string> aliases, IDictionary<string, string> globalProperties, CancellationToken cancellationToken)
-        {
-            // use loader service to determine output file for project if possible
-            string outputFilePath = null;
-
-            try
-            {
-                outputFilePath = await ProjectFileLoader.GetOutputFilePathAsync(projectFilePath, globalProperties, cancellationToken).ConfigureAwait(false);
-            }
-            catch (Exception e)
-            {
-                this.OnWorkspaceFailed(new WorkspaceDiagnostic(WorkspaceDiagnosticKind.Failure, e.Message));
-            }
-
-            if (outputFilePath != null && File.Exists(outputFilePath))
-            {
-                if (Workspace.TestHookStandaloneProjectsDoNotHoldReferences)
-                {
-                    var documentationService = this.Services.GetService<IDocumentationProviderService>();
-                    var docProvider = documentationService.GetDocumentationProvider(outputFilePath);
-                    var metadata = AssemblyMetadata.CreateFromImage(File.ReadAllBytes(outputFilePath));
-
-                    return metadata.GetReference(
-                        documentation: docProvider,
-                        aliases: aliases,
-                        display: outputFilePath);
-                }
-                else
-                {
-                    var metadataService = this.Services.GetService<IMetadataService>();
-                    return metadataService.GetReference(outputFilePath, new MetadataReferenceProperties(MetadataImageKind.Assembly, aliases));
-                }
-            }
-
-            return null;
-        }
         #endregion
 
         #region Apply Changes
@@ -1014,11 +254,11 @@ namespace Microsoft.CodeAnalysis.MSBuild
                 {
                     var projectPath = project.FilePath;
                     IProjectFileLoader loader;
-                    if (this.TryGetLoaderFromProjectPath(projectPath, ReportMode.Ignore, out loader))
+                    if (_loader.TryGetLoaderFromProjectPath(projectPath, out loader))
                     {
                         try
                         {
-                            _applyChangesProjectFile = loader.LoadProjectFileAsync(projectPath, _properties, CancellationToken.None).Result;
+                            _applyChangesProjectFile = loader.LoadProjectFileAsync(projectPath, _loader.Properties, CancellationToken.None).Result;
                         }
                         catch (System.IO.IOException exception)
                         {
@@ -1096,7 +336,7 @@ namespace Microsoft.CodeAnalysis.MSBuild
             var project = this.CurrentSolution.GetProject(info.Id.ProjectId);
 
             IProjectFileLoader loader;
-            if (this.TryGetLoaderFromProjectPath(project.FilePath, ReportMode.Ignore, out loader))
+            if (_loader.TryGetLoaderFromProjectPath(project.FilePath, out loader))
             {
                 var extension = _applyChangesProjectFile.GetDocumentExtension(info.SourceCodeKind);
                 var fileName = Path.ChangeExtension(info.Name, extension);
@@ -1255,5 +495,5 @@ namespace Microsoft.CodeAnalysis.MSBuild
             this.OnAnalyzerReferenceRemoved(projectId, analyzerReference);
         }
     }
-    #endregion
+#endregion
 }
