+++ src/Workspaces/Core/Desktop/Workspaces.Desktop.csproj
@@ -127,6 +127,7 @@
     <Compile Include="Workspace\MSBuild\SolutionFile\SolutionFile.cs" />
     <Compile Include="Workspace\MSBuild\VisualBasic\VisualBasicProjectFileLoader.cs" />
     <Compile Include="Workspace\MSBuild\VisualBasic\VisualBasicProjectFileLoaderFactory.cs" />
+    <Compile Include="Workspace\MSBuild\MSBuildProjectLoader.cs" />
   </ItemGroup>
   <ItemGroup>
     <InternalsVisibleTo Include="Microsoft.CodeAnalysis.CSharp.EditorFeatures" />
